requirements.txt
opencv-python
cvzone
pyttsx3
numpy
streamlit
ultralytics
pillow